Update to berkshelf 6.2.2 and install as isolated application
Recent changes to bundler mean that it doesn't appear to load the
gem path the same way as before, so berks wouldn't run without
bundle exec.

Use this as an opportunity to install berkshelf with a standalone
set of gems, rather than system gems, so that it doesn't affect
any of the installed chef etc on the system.