Update to use bento/ubuntu-18.04 base on vagrant
The ubuntu base box published by canonical for some reason hangs
during boot, without launching sshd, when launched in packer.
Switch over to the bento base instead.

This seems to have some conflicts with some of the default grub
/ dist package configs which were triggering interactive conflict
resolution and hanging the build, switch to a full non-interactive
dist-upgrade to fix that.